src/java/br/cesjf/lpwsd/Pagamento.java

package br.cesjf.lpwsd;

import java.util.Date;

public class Pagamento {
    private Aluno aluno;
    private Date dataPagamento;
    private double valor;

    
    public Aluno getAluno() {
        return aluno;
    }
    
    public void setAluno(Aluno aluno) {
        this.aluno = aluno;
    }
    
    public Date getDataPagamento() {
        return dataPagamento;
    }
    
    public void setDataPagamento(Date dataPagamento) {
        this.dataPagamento = dataPagamento;
    }
    
    public double getValor() {
        return valor;
    }
    
    public void setValor(double valor) {
        this.valor = valor;
    }
    
}